requirements.txt
transformers>=4.36.0
sentence-transformers>=2.2.2
faiss-cpu>=1.9.0
pdfplumber>=0.10.3
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.21
cryptography>=41.0.7
sqlalchemy>=2.0.23
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.3
python-multipart>=0.0.6
watchdog>=3.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
chromadb>=0.4.22
requests>=2.31.0
